Decode AionContractDetailsImpl from RLPContractDetails
The previous commit started the separation of the decoding process
into (a) reading the data from disk and (b) interpreting the data.
It also introduced part (a) as the process of creating the
RLPContractDetails object.

This commit completes the separation by doing part (b). As a result
only the interpretation of data remains inside the decode method of
AionContractDetailsImpl.

The unit tests for RLPContractDetails cover incorrect encoding size,
testEncodingIncorrectSize was no longer useful.