Perform data store creation in DetailsDataStore
DetailsDataStore has access to the physical databases where contract
data is stored. Since connecting the contract details object to the
data source must go though the DetailsDataStore it is best to perform
this operation when the instance is created.

Additionaly, this refactoring minimizes direct access to the databases
used by all contracts and reduces the mutability of the contract
details object.