Add AvmContractDetails using copy of current code
The new AvmContractDetails is identical to AionContractDetailsImpl,
other than the class name. This choice was made to allow easily tracking
the evolution of the class towards supporting only AVM contracts.
Similarly, AionContractDetailsImpl will evolve to support only the FVM.

The DetailsDataStore is responsible for instantiating the correct version
based on the given VM type. To avoid incorrect cross-class casting, it
does not create a ContractDetails instance when the VM type does not
correspond to a supported VM.