Remove storage transition from AvmContractDetails
The storage transition from in-line the details database to the external
storage database existed as a way to accomodate small FVM contracts that
could function mainly in the limited trie size that allows the in-line
storage. The AVM uses its object graph for tracking internal variables,
while the external storage is specifically meant for large collections
of data. Transitioning in this case is typically imminent and not worth
the additional cost demonstrated by the benchmark test.