Create staking blocks with given parent
The staking blocks were always created on top of the best block. This
is a valid setup for the staker tools but it prevents testing the
functionality with side chains.

Additionally, this is a bug fix for AionHub. The hub could cause incorrect
block creation due to the fact that it was calling the getBestBlock method
in the blockchain twice without synchronization.
This bug fix may relate to AKI-657.