Delete SpecialLockedDatabase
While it may offer minor optimizations when used for LevelDB and RocksDB
the class was not correctly maintained and lacks functionality. Further,
it is currently not necessary to use locked databases in production so
it is not worth keeping the class. It can be re-added with full correct
implementation when it is truly needed and proven to improve performance.